requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
Pillow==10.0.0